feat(442): inject project context on Agent dispatch (v4.18.2)
UserPromptSubmit hooks fire only on real user prompts. When the LLM
dispatches a subagent via the Agent tool, that subagent runs in an
isolated context and does NOT inherit the parent thread's
UserPromptSubmit-injected project context. Any orchestrator-style
ArcKit subagent whose body says "the ArcKit Project Context hook has
already detected all projects, artifacts…" silently lost that
context when invoked indirectly (general-purpose agent → Skill tool,
recursive Agent dispatch, etc.).

New hook arckit-claude/hooks/inject-agent-context.mjs:

- PreToolUse, matcher: Agent
- Skips when subagent_type doesn't start with arckit-* (no spam to
  Plan / Explore / general-purpose / claude-code-guide).
- Skips when subagent_type ends in -reader or -writer (the
  orchestrator pattern's strict-JSON tier — prepending prose context
  would pollute schema discipline).
- Otherwise builds the project-context block and prepends it to
  tool_input.prompt via updatedInput (the only PreToolUse return
  field that propagates into the dispatched subagent's context;
  additionalContext stays in the parent thread per
  https://code.claude.com/docs/en/hooks.md).

Project-scanning logic factored into a new shared
hooks/project-context-builder.mjs so the UserPromptSubmit hook
(arckit-context.mjs) and the new PreToolUse hook produce identical
output. Verified with 6 unit cases: arckit-research → inject;
arckit-grants-reader / arckit-datascout-writer → skip;
general-purpose → skip; non-Agent tool → skip; no projects/ dir →
silent pass-through.
