+++ .claude-plugin/marketplace.json
@@ -13,7 +13,7 @@
       "name": "arckit",
       "source": "./arckit-claude",
       "description": "70 slash commands for enterprise architecture artifacts, vendor procurement, and UK Government compliance",
-      "version": "4.18.1",
+      "version": "4.18.2",
       "author": {
         "name": "TractorJuice"
       },
