+++ README.md
@@ -108,7 +108,7 @@ uv tool install arckit-cli --from git+https://github.com/tractorjuice/arc-kit.gi
 uvx --from git+https://github.com/tractorjuice/arc-kit.git arckit init my-project
 ```
 
-**Latest Release**: [v4.18.1](https://github.com/tractorjuice/arc-kit/releases/tag/v4.18.1)
+**Latest Release**: [v4.18.2](https://github.com/tractorjuice/arc-kit/releases/tag/v4.18.2)
 
 ### Platform Support
 
@@ -1708,7 +1708,7 @@ arckit init .
 
 - **Issues**: [GitHub Issues](https://github.com/tractorjuice/arc-kit/issues)
 - **Releases**: [GitHub Releases](https://github.com/tractorjuice/arc-kit/releases)
-- **Latest Version**: [v4.18.1](https://github.com/tractorjuice/arc-kit/releases/tag/v4.18.1)
+- **Latest Version**: [v4.18.2](https://github.com/tractorjuice/arc-kit/releases/tag/v4.18.2)
 
 ---
