+++ arckit-claude/.claude-plugin/plugin.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://json.schemastore.org/claude-code-plugin-manifest.json",
   "name": "arckit",
-  "version": "4.18.1",
+  "version": "4.18.2",
   "description": "Enterprise Architecture Governance & Vendor Procurement Toolkit - 70 slash commands for generating architecture artifacts",
   "author": {
     "name": "TractorJuice",
