+++ arckit-gemini/gemini-extension.json
@@ -1,6 +1,6 @@
 {
   "name": "arckit",
-  "version": "4.18.1",
+  "version": "4.18.2",
   "description": "Enterprise Architecture Governance & Vendor Procurement Toolkit - 68 commands for architecture artifacts, vendor procurement, and UK Government compliance",
   "contextFileName": "GEMINI.md",
   "mcpServers": {
