+++ arckit-paperclip/VERSION
@@ -1,1 +1,1 @@
-4.18.1
+4.18.2
