+++ arckit-paperclip/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tractorjuice/arckit-paperclip",
-  "version": "4.18.1",
+  "version": "4.18.2",
   "description": "ArcKit Enterprise Architecture Governance plugin for Paperclip — 68 commands for generating architecture artifacts",
   "type": "module",
   "main": "dist/worker.js",
