+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "arckit-cli"
-version = "4.18.1"
+version = "4.18.2"
 description = "ArcKit CLI - Enterprise Architecture Governance & Vendor Procurement Toolkit"
 requires-python = ">=3.11"
 dependencies = [
